Clean up more routes
